--- .idx/dev.nix
@@ -1,26 +1,0 @@
-{pkgs}: {
-  channel = "stable-24.05";
-  packages = [
-    pkgs.nodejs_20
-  ];
-  idx.extensions = [
-    
-  ];
-  idx.previews = {
-    previews = {
-      web = {
-        command = [
-          "npm"
-          "run"
-          "dev"
-          "--"
-          "--port"
-          "$PORT"
-          "--hostname"
-          "0.0.0.0"
-        ];
-        manager = "web";
-      };
-    };
-  };
-}
